+++ olympics-website/prisma/schema.prisma
@@ -62,6 +62,8 @@ model Score {
   score        Float
   team         Team       @relation(fields: [teamId], references: [id])
   teamId       Int
+
+  @@unique([teamId, eventGamesId])
 }
